refactor: extract custom provider transport and tool streaming
